feat: quantité Main d'Oeuvre en décimal (pas de 0.25) dans forfait et prestation
Changement du type int vers double dans les entités backend et
modification des InputNumber pour accepter des valeurs décimales.

+++ backend/src/main/java/net/nanthrax/moussaillon/persistence/ForfaitMainOeuvreEntity.java
@@ -10,5 +10,5 @@ public class ForfaitMainOeuvreEntity extends PanacheEntity {
     @ManyToOne
     public MainOeuvreEntity mainOeuvre;
 
-    public int quantite;
+    public double quantite;
 }
