+++ backend/src/main/java/net/nanthrax/moussaillon/persistence/ServiceMainOeuvreEntity.java
@@ -10,5 +10,5 @@ public class ServiceMainOeuvreEntity extends PanacheEntity {
     @ManyToOne
     public MainOeuvreEntity mainOeuvre;
 
-    public int quantite;
+    public double quantite;
 }
